+++ benchmark/comparison_benchmark.rb
@@ -37,15 +37,45 @@ begin
   Benchmark.ips do |x|
     x.config(time: 10, warmup: 5)
 
-    x.report("OSV - Hash output") { File.open("benchmark/test.csv") { |f| OSV.for_each(f).to_a } }
+    x.report("OSV - Hash output") do
+      result = []
+      File.open("benchmark/test.csv") { |f| OSV.for_each(f) { |row| result << row } }
+      result
+    end
 
-    x.report("CSV - Hash output") { File.open("benchmark/test.csv") { |f| CSV.new(f, headers: true).map(&:to_h) } }
+    x.report("OSV - Array output") do
+      result = []
+      File.open("benchmark/test.csv") { |f| OSV.for_each(f, result_type: :array) { |row| result << row } }
+      result
+    end
 
-    x.report("OSV - Array output") { File.open("benchmark/test.csv") { |f| OSV.for_each(f, result_type: :array).to_a } }
+    x.report("OSV - Direct Open Array output") do
+      result = []
+      OSV.for_each("benchmark/test.csv", result_type: :array) { |row| result << row }
+      result
+    end
 
-    x.report("OSV - Direct Open Array output") { |f| OSV.for_each("benchmark/test.csv", result_type: :array).to_a }
+    x.report("OSV - StringIO") do
+      io = StringIO.new(test_data)
+      result = []
+      OSV.for_each(io) { |row| result << row }
+      result
+      io.close
+    end
 
-    x.report("CSV - Array output") { File.open("benchmark/test.csv") { |f| CSV.new(f).read } }
+    x.report("OSV - Gzipped") do
+      result = []
+      Zlib::GzipReader.open("benchmark/test.csv.gz") do |gz|
+        OSV.for_each(gz, result_type: :array) { |row| result << row }
+      end
+      result
+    end
+
+    x.report("OSV - Gzipped Direct") do
+      result = []
+      OSV.for_each("benchmark/test.csv.gz", result_type: :array) { |row| result << row }
+      result
+    end
 
     x.report("FastCSV - Array output") do
       result = []
@@ -53,17 +83,9 @@ begin
       result
     end
 
-    x.report("OSV - StringIO") do
-      io = StringIO.new(test_data)
-      OSV.for_each(io).to_a
-      io.close
-    end
-
-    x.report("CSV - StringIO") do
-      io = StringIO.new(test_data)
-      result = CSV.new(io, headers: true).map(&:to_h)
-      io.close
-
+    x.report("FastCSV - Gzipped") do
+      result = []
+      Zlib::GzipReader.open("benchmark/test.csv.gz") { |gz| FastCSV.raw_parse(gz) { |row| result << row } }
       result
     end
 
@@ -76,11 +98,21 @@ begin
       result
     end
 
-    x.report("OSV - Gzipped") { Zlib::GzipReader.open("benchmark/test.csv.gz") { |gz| OSV.for_each(gz).to_a } }
+    # x.report("CSV - Gzipped") do
+    #   Zlib::GzipReader.open("benchmark/test.csv.gz") { |gz| CSV.new(gz, headers: true).map(&:to_h) }
+    # end
 
-    x.report("CSV - Gzipped") do
-      Zlib::GzipReader.open("benchmark/test.csv.gz") { |gz| CSV.new(gz, headers: true).map(&:to_h) }
-    end
+    # x.report("CSV - Hash output") { File.open("benchmark/test.csv") { |f| CSV.new(f, headers: true).map(&:to_h) } }
+
+    # x.report("CSV - StringIO") do
+    #   io = StringIO.new(test_data)
+    #   result = CSV.new(io, headers: true).map(&:to_h)
+    #   io.close
+
+    #   result
+    # end
+
+    # x.report("CSV - Array output") { File.open("benchmark/test.csv") { |f| CSV.new(f).read } }
 
     x.compare!
   end